Updated to 1.3.1.
Updated ModSettings.getMcinst() to just use Minecraft.getMinecraft() now. Also, because of this got rid of ModSettings.presetMcint(Minecraft) because it's not like anyone used it, and I honestly have no idea why the hell it was there in the first place tbh because it would just break stuff.
Removed an unused import for WidgetItem2DRender, which shouldn't have been there anyways (It was a leftover from debugging stuff).
WidgetItem2DRender now has to use reflection to get the isDrawing field from the tesselator because mojang made it a private field. If it can't get it, it will just print to console and that part of the protection will be removed, but it's not likely to happen so.