Major update. 0.15.2.
Repackaged everything to sharose.mods.guiapi.

Examples are also repackaged to sharose.mods.guiapi.examples.
The GuiAPI mod file is now a forge style mod.
GuiAPI as a whole requires Forge now.
Fixed a bug relating to injecting the global mod options button (If you resized the window it would disappear and not come back)
Made GuiAPI a coremod: It's just a dummy and doesn't do anything, but it will have TWL and such injected into the classpath far earlier than a regular mod, removing (for good) the need to edit the jar.