Updated to 0.15.3 for 1.5.
Also, the button add code is hacked together now so it forces it into the corner between touchscreen mode and difficulty because there's literally nowhere else to stick it. I'm so sorry.

I hope I can get around to rewriting the internals later so they are fit for forge, since apparently whoever it was who asked for it didn't want to put in the effort for it and thus the deadline was missed. At least when I do it it'll hopefully use coremodding and stuff on TWL so I don't need to manually patch it all whenever I get around to updating it. Also it'll be less shit internally, maybe.